+++ src/main/java/com/hkbu/comp7640/service/TransactionService.java
@@ -20,6 +20,6 @@ public interface TransactionService extends IService<Transaction> {
      * @param page 分页
      * @return
      */
-    IPage<TransactionWithProductVendorDTO> pageTransaction(Page<TransactionWithProductVendorDTO> page, Long userId);
+    IPage<TransactionWithProductVendorDTO> pageTransaction(Page<TransactionWithProductVendorDTO> page, Integer userId);
 
 }
